Move Pubkey into its own module
