+++ benches/banking_stage.rs
@@ -12,7 +12,8 @@ use solana::banking_stage::{BankingStage, NUM_THREADS};
 use solana::entry::Entry;
 use solana::mint::Mint;
 use solana::packet::{to_packets_chunked, PacketRecycler};
-use solana::signature::{KeypairUtil, Pubkey, Signature};
+use solana::pubkey::Pubkey;
+use solana::signature::{KeypairUtil, Signature};
 use solana::system_transaction::SystemTransaction;
 use solana::transaction::Transaction;
 use std::iter;
