+++ src/bank.rs
@@ -17,7 +17,8 @@ use ledger::Block;
 use log::Level;
 use mint::Mint;
 use payment_plan::Payment;
-use signature::{Keypair, Pubkey, Signature};
+use pubkey::Pubkey;
+use signature::{Keypair, Signature};
 use std;
 use std::collections::{BTreeMap, HashMap, VecDeque};
 use std::result;
