+++ src/broadcast_stage.rs
@@ -283,8 +283,9 @@ mod tests {
     use entry::Entry;
     use ledger::next_entries_mut;
     use mint::Mint;
+    use pubkey::Pubkey;
     use service::Service;
-    use signature::{Keypair, KeypairUtil, Pubkey};
+    use signature::{Keypair, KeypairUtil};
     use std::cmp;
     use std::sync::atomic::AtomicBool;
     use std::sync::mpsc::{channel, Sender};
