+++ src/crdt.rs
@@ -21,10 +21,11 @@ use ledger::LedgerWindow;
 use log::Level;
 use netutil::{bind_in_range, bind_to, multi_bind_in_range};
 use packet::{to_blob, Blob, BlobRecycler, SharedBlob, BLOB_SIZE};
+use pubkey::Pubkey;
 use rand::{thread_rng, Rng};
 use rayon::prelude::*;
 use result::{Error, Result};
-use signature::{Keypair, KeypairUtil, Pubkey};
+use signature::{Keypair, KeypairUtil};
 use std;
 use std::collections::HashMap;
 use std::net::{IpAddr, Ipv4Addr, SocketAddr, UdpSocket};
@@ -1408,8 +1409,9 @@ mod tests {
     use ledger::{LedgerWindow, LedgerWriter};
     use logger;
     use packet::BlobRecycler;
+    use pubkey::Pubkey;
     use result::Error;
-    use signature::{Keypair, KeypairUtil, Pubkey};
+    use signature::{Keypair, KeypairUtil};
     use std::fs::remove_dir_all;
     use std::net::{IpAddr, Ipv4Addr, SocketAddr};
     use std::sync::atomic::{AtomicBool, Ordering};
