+++ src/drone.rs
@@ -8,8 +8,8 @@ use bincode::{deserialize, serialize};
 use bytes::Bytes;
 use influx_db_client as influxdb;
 use metrics;
-use signature::Signature;
-use signature::{Keypair, Pubkey};
+use pubkey::Pubkey;
+use signature::{Keypair, Signature};
 use std::io;
 use std::io::{Error, ErrorKind};
 use std::net::{IpAddr, Ipv4Addr, SocketAddr, UdpSocket};
