+++ src/entry.rs
@@ -7,8 +7,8 @@ use budget_transaction::BudgetTransaction;
 use hash::Hash;
 use packet::{BlobRecycler, SharedBlob, BLOB_DATA_SIZE};
 use poh::Poh;
+use pubkey::Pubkey;
 use rayon::prelude::*;
-use signature::Pubkey;
 use std::io::Cursor;
 use std::net::SocketAddr;
 use std::sync::mpsc::{Receiver, Sender};
