+++ src/fullnode.rs
@@ -7,10 +7,11 @@ use drone::DRONE_PORT;
 use entry::Entry;
 use ledger::read_ledger;
 use ncp::Ncp;
+use pubkey::Pubkey;
 use rpc::{JsonRpcService, RPC_PORT};
 use rpu::Rpu;
 use service::Service;
-use signature::{Keypair, KeypairUtil, Pubkey};
+use signature::{Keypair, KeypairUtil};
 use std::net::UdpSocket;
 use std::net::{IpAddr, Ipv4Addr, SocketAddr};
 use std::sync::atomic::{AtomicBool, Ordering};
