+++ src/packet.rs
@@ -7,11 +7,11 @@ use hash::Hash;
 #[cfg(test)]
 use ledger::{next_entries_mut, Block};
 use log::Level;
+use pubkey::Pubkey;
 use recvmmsg::{recv_mmsg, NUM_RCVMMSGS};
 use recycler;
 use result::{Error, Result};
 use serde::Serialize;
-use signature::Pubkey;
 use std::fmt;
 use std::io;
 use std::mem::size_of;
