+++ src/rpc.rs
@@ -6,8 +6,9 @@ use bincode::deserialize;
 use bs58;
 use jsonrpc_core::*;
 use jsonrpc_http_server::*;
+use pubkey::Pubkey;
 use service::Service;
-use signature::{Pubkey, Signature};
+use signature::Signature;
 use std::mem;
 use std::net::{SocketAddr, UdpSocket};
 use std::result;
