+++ src/thin_client.rs
@@ -10,9 +10,10 @@ use crdt::{Crdt, CrdtError, NodeInfo};
 use hash::Hash;
 use log::Level;
 use ncp::Ncp;
+use pubkey::Pubkey;
 use request::{Request, Response};
 use result::{Error, Result};
-use signature::{Keypair, Pubkey, Signature};
+use signature::{Keypair, Signature};
 use std;
 use std::collections::HashMap;
 use std::io;
