+++ src/vote_stage.rs
@@ -10,8 +10,9 @@ use influx_db_client as influxdb;
 use log::Level;
 use metrics;
 use packet::{BlobRecycler, SharedBlob};
+use pubkey::Pubkey;
 use result::Result;
-use signature::{Keypair, Pubkey};
+use signature::Keypair;
 use std::result;
 use std::sync::atomic::AtomicUsize;
 use std::sync::{Arc, RwLock};
