+++ src/wallet.rs
@@ -8,12 +8,13 @@ use crdt::NodeInfo;
 use drone::DroneRequest;
 use fullnode::Config;
 use hash::Hash;
+use pubkey::Pubkey;
 use reqwest;
 use reqwest::header::CONTENT_TYPE;
 use ring::rand::SystemRandom;
 use ring::signature::Ed25519KeyPair;
 use serde_json::{self, Value};
-use signature::{Keypair, KeypairUtil, Pubkey, Signature};
+use signature::{Keypair, KeypairUtil, Signature};
 use std::fs::{self, File};
 use std::io::prelude::*;
 use std::io::{Error, ErrorKind, Write};
