+++ src/window_service.rs
@@ -5,9 +5,9 @@ use crdt::{Crdt, NodeInfo};
 use entry::EntrySender;
 use log::Level;
 use packet::{BlobRecycler, SharedBlob};
+use pubkey::Pubkey;
 use rand::{thread_rng, Rng};
 use result::{Error, Result};
-use signature::Pubkey;
 use std::net::UdpSocket;
 use std::sync::atomic::{AtomicBool, AtomicUsize, Ordering};
 use std::sync::mpsc::RecvTimeoutError;
