+++ tests/multinode.rs
@@ -13,9 +13,10 @@ use solana::ledger::{read_ledger, LedgerWriter};
 use solana::logger;
 use solana::mint::Mint;
 use solana::ncp::Ncp;
+use solana::pubkey::Pubkey;
 use solana::result;
 use solana::service::Service;
-use solana::signature::{Keypair, KeypairUtil, Pubkey};
+use solana::signature::{Keypair, KeypairUtil};
 use solana::thin_client::ThinClient;
 use solana::timing::{duration_as_ms, duration_as_s};
 use solana::window::{default_window, WINDOW_SIZE};
